Update at night 19/04/2020 10:00PM

+++ DanhGiaDoanVien/DTO/GoodStudent.cs
@@ -12,13 +12,13 @@ namespace DanhGiaDoanVien.NewFolder1
         private int id;
         private string idStudent;
         private string idSemester;
-        private byte vote;
-        private byte totalVote;
+        private int vote;
+        private int totalVote;
         public int Id { get => id; set => id = value; }
         public string IdStudent { get => idStudent; set => idStudent = value; }
         public string IdSemester { get => idSemester; set => idSemester = value; }
-        public byte Vote { get => vote; set => vote = value; }
-        public byte TotalVote { get => totalVote; set => totalVote = value; }
+        public int Vote { get => vote; set => vote = value; }
+        public int TotalVote { get => totalVote; set => totalVote = value; }
 
         public GoodStudent() { }
 
@@ -27,8 +27,8 @@ namespace DanhGiaDoanVien.NewFolder1
             this.Id = (int)row["id"];
             this.IdStudent = (string)row["MSSV"];
             this.IdSemester = (string)row["idNamHoc"];
-            this.Vote = (byte)row["phieuBau"];
-            this.TotalVote = (byte)row["tongPhieu"];
+            this.Vote = (int)row["phieuBau"];
+            this.TotalVote = (int)row["tongPhieu"];
         }
     }
 }
