+++ DanhGiaDoanVien/DTO/Group.cs
@@ -11,23 +11,23 @@ namespace DanhGiaDoanVien.DTO
     {
         private string id;
         private string name;
-        private byte amountMembet;
-        private byte totalStudent;
-        private byte totalTeacher;
+        private int amountMembet;
+        private int totalStudent;
+        private int totalTeacher;
         public string Id { get => id; set => id = value; }
         public string Name { get => name; set => name = value; }
-        public byte AmountMembet { get => amountMembet; set => amountMembet = value; }
-        public byte TotalStudent { get => totalStudent; set => totalStudent = value; }
-        public byte TotalTeacher { get => totalTeacher; set => totalTeacher = value; }
+        public int AmountMembet { get => amountMembet; set => amountMembet = value; }
+        public int TotalStudent { get => totalStudent; set => totalStudent = value; }
+        public int TotalTeacher { get => totalTeacher; set => totalTeacher = value; }
 
         public Group() { }
         public Group(DataRow row)
         {
             Id = (string)row["id"];
             Name = (string)row["ten"];
-            AmountMembet = (byte)row["soDoanVien"];
-            TotalStudent = (byte)row["tongSV"];
-            totalTeacher = (byte)row["tongGV"];
+            AmountMembet = (int)row["soDoanVien"];
+            TotalStudent = (int)row["tongSV"];
+            totalTeacher = (int)row["tongGV"];
         }
     }
 }
