+++ DanhGiaDoanVien/DTO/ScoresStudent.cs
@@ -15,9 +15,9 @@ namespace DanhGiaDoanVien.DTO
         private float averageSemester1;
         private float averageSemester2;
         private float totalAverage;
-        private byte pointTraning1;
-        private byte pointTraning2;
-        private byte averageTraningPoint;
+        private int pointTraning1;
+        private int pointTraning2;
+        private int averageTraningPoint;
         private string rank;
         private string achievement;
         private bool isGoodMember;
@@ -30,9 +30,9 @@ namespace DanhGiaDoanVien.DTO
         public float AverageSemester1 { get => averageSemester1; set => averageSemester1 = value; }
         public float AverageSemester2 { get => averageSemester2; set => averageSemester2 = value; }
         public float TotalAverage { get => totalAverage; set => totalAverage = value; }
-        public byte PointTraning1 { get => pointTraning1; set => pointTraning1 = value; }
-        public byte PointTraning2 { get => pointTraning2; set => pointTraning2 = value; }
-        public byte AverageTrainingPoint { get => averageTraningPoint; set => averageTraningPoint = value; }
+        public int PointTraning1 { get => pointTraning1; set => pointTraning1 = value; }
+        public int PointTraning2 { get => pointTraning2; set => pointTraning2 = value; }
+        public int AverageTrainingPoint { get => averageTraningPoint; set => averageTraningPoint = value; }
         public string Rank { get => rank; set => rank = value; }
         public string Achievement { get => achievement; set => achievement = value; }
         public bool IsGoodMember { get => isGoodMember; set => isGoodMember = value; }
@@ -48,9 +48,9 @@ namespace DanhGiaDoanVien.DTO
             AverageSemester1 = (float)row["diemHK1"];
             AverageSemester2 = (float)row["diemHK2"];
             TotalAverage = (float)row["tongHK"];
-            PointTraning1 = (byte)row["DRLHK1"];
-            PointTraning2 = (byte)row["DRLHK2"];
-            AverageTrainingPoint = (byte)row["DRL"];
+            PointTraning1 = (int)row["DRLHK1"];
+            PointTraning2 = (int)row["DRLHK2"];
+            AverageTrainingPoint = (int)row["DRL"];
             Rank = (string)row["xepLoai"];
             Achievement = (string)row["thanhTichTieuBieu"];
             Note = (string)row["ghiChu"];
